Extract types from api store
