+++ dashboard/src/helpers/subscriptions.ts
@@ -2,7 +2,7 @@ import banner1 from "assets/subscriptions/subscription_banner_1.png";
 import banner2 from "assets/subscriptions/subscription_banner_2.png";
 import banner3 from "assets/subscriptions/subscription_banner_3.png";
 import banner4 from "assets/subscriptions/subscription_banner_4.png";
-import { Subscription, SubscriptionPrice } from "store/api";
+import { Subscription, SubscriptionPrice } from "types";
 
 const banners = [
   {
