+++ dashboard/src/store/index.ts
@@ -1,16 +1,16 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { apiSlice } from "./api";
+import { api } from "./api";
 import settingsReducer from "./settings";
 import trackReducer from "./track";
 
 export const store = configureStore({
   reducer: {
-    [apiSlice.reducerPath]: apiSlice.reducer,
+    [api.reducerPath]: api.reducer,
     settings: settingsReducer,
     track: trackReducer,
   },
   middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware().concat(apiSlice.middleware),
+    getDefaultMiddleware().concat(api.middleware),
 });
 
 export type RootState = ReturnType<typeof store.getState>;
